+++ tests/testthat/test-ghcnd.R
@@ -35,6 +35,8 @@ test_that("version metadata", {
 
 test_that("search for data", {
   skip_on_cran()
+  skip_on_travis()
+  skip_on_appveyor()
   skip_if_government_down()
 
   search_a <- ghcnd_search("AGE00147704", var = "PRCP")
@@ -58,6 +60,8 @@ test_that("search for data", {
 
 test_that("get data", {
   skip_on_cran()
+  skip_on_travis()
+  skip_on_appveyor()
   skip_if_government_down()
 
   aa <- ghcnd(stationid = "AGE00147704")