+++ tests/testthat/test-tornadoes.R
@@ -4,6 +4,6 @@ test_that("tornadoes works", {
   skip_on_cran()
   skip_if_government_down()
 
-  expect_warning(torn <- tornadoes(), regexp = "Dropping null geometries")
+  torn <- tornadoes()
   expect_is(torn, "SpatialLinesDataFrame")
 })