+++ tests/testthat/test-vis_miss.R
@@ -1,8 +1,12 @@
 context("vis_miss")
 
 test_that("Valid ggplot object is produced",{
+  skip_on_cran()
+  skip_on_travis()
+  skip_on_appveyor()
+
   monitors <- c("ASN00003003", "ASM00094299")
-  weather_df <- meteo_pull_monitors(monitors)
+  weather_df <- suppressMessages(meteo_pull_monitors(monitors))
   out <- vis_miss(weather_df)
 
   expect_is(out, "ggplot")